requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
pdfrw==0.4
pillow==10.4.0
reportlab==4.2.5
sqlparse==0.5.1
tzdata==2024.2
